rag-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1
numpy>=2.0.0
pydantic==2.10.4
httpx==0.28.1
